ui tab switching works now!1
* Now the 4th iteration on the layout switching. Creating a complete new
layout, populating it and then switching the main layout to it works. Qt
manages deletion in the background. Deleting yourself does not work so
easily.
* I added a sketch of a db unittest which handles creating/dropping a
new table with a "test string" name (with spaces) and setting/getting
something. Still not finished there but could start working with it. I
btw: found how to handle the log correctly :]