editing and submitting configs done!
One can now edit the description info for a module, author, date and
version and submit it all to the json in the module folder. A new python
module a2ctrl was created to host variable Qt compound controls with a
standardized interface. Also there is a new
checkbox-modName-version-author-header which changes on selection change
:)