ui is QMainWindow now, simpler config scheme
I finally managed to have the ui design in a "MainWindow"-type file. It
does not have to be edited anymore to be 'correct' it can totally be
imported right from pyside uic!
And now I simplified the config.json scheme to be a list with dicts
only! First element is always the nfo part. This clears up a lot of
handling code. elements to handle the nfo typ will then be in a2ctrl.py
... WIP! Submitting is broken now.