first cfg element added dynamically
ok. One can now add "include" controls in the cfg. It's not yet
functional. so nothing is build but the edit, draw structure seems not
much more straightforward. I need to solve the getCfg vs tempConfig
thing though... thing is: I could not add widgets to the mainLayout...
So I probably just add stuff to the inflight cfg and draw the edit ctrls
again. But I need to get changes of all the ctrls before or stuff could
be forgotten. But then if I implement this and have an airtight
tempConfig I don't need to get a complete config in the end. That's
actually nice as well....